Move types and functions related to freezing to skjson
The basic definitions are currently in
skiplang/skjson/ts/binding/src/index.ts with other definitions built on them
in skipruntime-ts/core/src/index.ts, skipruntime-ts/api/src/internals.ts,
and skipruntime-ts/api/src/api.ts. This PR moves all these other definitions
to skjson with the basic ones. In addition to placing code that is tightly
interdependent close together, this allows them to be surfaced in the
@skipruntime/api package without making it depend on @skipruntime/core.
